pyproject.toml
[project]
name = "ixlab-sshcli"
version = "1.0.0"
description = "Command-line explorer for SSH configs built on sshcore."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Iacopo Palazzi", email = "iacopo.palazzi@gmail.com" },
]
license = { text = "MIT" }
keywords = ["ssh", "cli", "config"]
dependencies = [
  "ixlab-sshcore>=1.0.0",
  "typer>=0.12",
  "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/iakko/sshcli"

[project.scripts]
sshcli = "sshcli:run"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sshcli*"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
